Add responsive web layout and refactor code
