+++ lib/utils/global.dart
@@ -1,0 +1,4 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
+
+final FirebaseFirestore firestore = FirebaseFirestore.instance;
+const webScreenSize = 600;
