+++ pubspec.yaml
@@ -34,7 +34,8 @@ dependencies:
   cloud_firestore: ^3.1.11
   cupertino_icons: ^1.0.2
   syncfusion_flutter_charts: ^20.1.47
-  rolling_switch: ^0.0.2
+  # rolling_switch: ^0.0.2
+  responsive_builder: ^0.4.2
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
